pyproject.toml
[project]
name = "shipgit"
version = "0.1.0"
description = "A lightweight Git shipping assistant for safe commits, tags, and multi-remote publishing."
requires-python = ">=3.10"
dependencies = [
    "typer",
    "rich",
    "questionary",
    "pyyaml"
]

[project.scripts]
shipgit = "shipgit.main:app"

[tool.ruff]
line-length = 100
target-version = "py311"
